updated ui and types
